# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Transaction",
        "version": "0.1"
    },
    "paths": {
        "/transaction-backend-exemple/public/api/transaction": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get list of transaction",
                "description": "Returns list of transaction",
                "operationId": "63c8786026df638efc415ae8364b5229",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Transaction(s) not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create a transaction",
                "description": "Create a transaction",
                "operationId": "170d1a85a011285db48279a1ae35968b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Transaction(s) not found"
                    },
                    "403": {
                        "description": "Bad Request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/transaction-backend-exemple/public/api/transaction/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get a transaction",
                "description": "Get a transaction",
                "operationId": "8b79d75df0c3aeb6776c084dc96e54c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Get a transaction",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Transaction(s) not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update a transaction",
                "description": "Update a transaction",
                "operationId": "e12f05d1d52aa3cf65e3cf3c5baac1de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to update transaction",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Transaction(s) not found"
                    },
                    "403": {
                        "description": "Bad Request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete a transaction",
                "description": "Delete a transaction",
                "operationId": "91c0ee680eb325898aebb852e04bdf69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to delete transaction",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/transaction-backend-exemple/public/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list user",
                "description": "Returns list user",
                "operationId": "18db51d4fb7da96668f7f91416084456",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User(s) not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a user",
                "description": "Create a user",
                "operationId": "d13ff856ff3d234cefb85a50a37f6b7f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User(s) not found"
                    },
                    "403": {
                        "description": "Bad Request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/transaction-backend-exemple/public/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get a user",
                "description": "Get a user",
                "operationId": "6d15ffc3818ad2b8fa2f113d35c9f2c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Get a user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User(s) not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update a user",
                "description": "Update a user",
                "operationId": "5dbd8d0aacea32f54c65ad51bf25389b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to update user",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User(s) not found"
                    },
                    "403": {
                        "description": "Bad Request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a user",
                "description": "Delete a user",
                "operationId": "633b0c3506fde7117a2f261149ce170e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to delete user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}